+++ src/tasks/utils/tasks.test.ts
@@ -1,11 +1,20 @@
 import { generateStoryData } from "../../KonvaTimeline/stories-data";
 import { InternalTimeRange } from "../../utils/time";
 
-import { validateTasks } from "./tasks";
+import { getTaskYCoordinate, validateTasks } from "./tasks";
 
 // From: Sunday, 1 January 2023 00:00:00 GMT+01:00
 // To: Monday, 2 January 2023 00:00:00 GMT+01:00
 const range: InternalTimeRange = { start: 1672527600000, end: 1672614000000 };
+
+describe("getTaskYCoordinate", () => {
+  it("valid", () => {
+    const ROW_HEIGHT = 50;
+    const resourceIndex = Math.ceil(Math.random() * 10);
+    const yCoordinate = getTaskYCoordinate(resourceIndex, ROW_HEIGHT);
+    expect(yCoordinate % ROW_HEIGHT).toEqual(ROW_HEIGHT * 0.1);
+  });
+});
 
 describe("validateTasks", () => {
   it("empty", () => {